min and max values change to double type and are now a 10 base log value in GUI

+++ Holovibes/main_window.hh
@@ -1,6 +1,7 @@
 #ifndef MAIN_WINDOW_HH_
 # define MAIN_WINDOW_HH_
 
+# include <cmath>
 # include <thread>
 # include <QMainWindow>
 # include <QFileDialog>
@@ -50,8 +51,8 @@ namespace gui
     void set_view_mode(QString value);
     void set_contrast_mode(bool value);
     void set_auto_contrast();
-    void set_contrast_min(int value);
-    void set_contrast_max(int value);
+    void set_contrast_min(double value);
+    void set_contrast_max(double value);
     void set_log_scale(bool value);
     void set_shifted_corners(bool value);
